Added ClaimbuildApplicationRepositoryImpl
